chore: rename crate to bls_bulletproofs
Yesterday we renamed this crate from blst_bulletproofs to sn_bulletproofs, but later we decided it
would be independent of Safe, so we'll rename again. We lose the 'sn' prefix in favour of 'bls',
which drops the 't' from the original name, since that's a specific implementation detail of the
original repository, which our fork has now diverged from. We also have other repos/crates that use
this 'bls' prefix, so it brings it in line with those.