ci: use standard pr and release workflows
Removes the previous workflows and adds the standard PR and release workflows applied to the other
repositories in the `maidsafe` organisation.

At the moment, udeps and clippy are not enabled. Fixing the problems with them either requires
knowledge of the library or significant investigation.

We also add the license verification tool.