chore: use license verification in pr workflow
As a team we had agreed that any repositories not prefixed with `sn_` would use the BSD-3-Clause
license; however, since this repository was forked, we're going to stick with the upstream MIT
license.

To get the license verification to pass, we have:
* Removed LICENSE.txt in favour of a LICENSE file
* Added a reference to the license type and the LICENSE file in the README
* Applied a copyright notice to every Rust source file

Note: the copyright notice is intentionally minimal and directs to the LICENSE file for full details.