chore: upgrade blstrs to 0.6.1
This should enable upgrading `blsttc` to 8.0.1 in `sn_dbc`. The `blsttc` crate has upgraded `blstrs`
to 0.6.1, so this gets both crates on the same version.

The `bls_ringct` crate can then upgrade to this new version of `bls_bulletproofs` and also be on the
same version of `blstrs`. This will mean there will only be one version of `blstrs` in the
dependency graph.

It was necessary to upgrade `group` to 0.12.0 as part of this change.